FCT-16897: refactor Avatar component
