+++ lib/components/Blocks/Header/index.stories.tsx
@@ -8,7 +8,7 @@ const meta = {
     title: "Alba Horneros",
     subtitle: "Product Designer",
     src: "https://github.com/dani-moreno.png",
-    alt: "@dani-moreno",
+    alt: "DM",
   },
 } satisfies Meta<typeof Header>
 
@@ -20,6 +20,6 @@ export const Default: Story = {
     title: "Alba Horneros",
     subtitle: "Product Designer",
     src: "https://github.com/dani-moreno.png",
-    alt: "@dani-moreno",
+    alt: "DM",
   },
 }
