+++ lib/components/Blocks/Header/index.tsx
@@ -1,5 +1,4 @@
-import { FlexBox } from "@/components/Layout/_FlexBox"
-import { Avatar, AvatarFallback, AvatarImage } from "@/ui/avatar"
+import { Avatar } from "@/components/Information/Avatar"
 import { forwardRef } from "react"
 
 interface HeaderProps {
@@ -12,20 +11,17 @@ interface HeaderProps {
 export const Header = forwardRef<HTMLDivElement, HeaderProps>(
   ({ title, subtitle, src, alt }, ref) => {
     return (
-      <FlexBox ref={ref} className="px-10 py-6">
-        <Avatar size="xlarge">
-          <AvatarImage src={src} alt={alt} />
-          <AvatarFallback>{alt}</AvatarFallback>
-        </Avatar>
-        <FlexBox className="flex-col gap-4 pl-5">
+      <div className="flex px-10 py-6" ref={ref}>
+        <Avatar size="xlarge" src={src} alt={alt} />
+        <div className="flex flex-col gap-4 pl-5">
           <h1 className="pt-2 text-[1.69rem]/[1.6rem] font-medium text-foreground">
             {title}
           </h1>
           <h2 className="text-[1.19rem]/4 font-normal text-intermediate">
             {subtitle}
           </h2>
-        </FlexBox>
-      </FlexBox>
+        </div>
+      </div>
     )
   }
 )
